Move system struct to system.h
